Allocate bytes, don't depend on shared underlying array
Turns out the emitted tokens are pointing to the same undelying array from bytes.Buffer. This is fine if the caller only stores one token at a time, as in typical calls to Scan.

If the caller puts several tokens into a collection, say, then they each token hold a reference back to that underlying bytes array, and they clobber each other.

It's shared global state, which I had a feeling about, but couldn't put a finger on. Discovered in jargon, which uses slices.

Downside is that this solution allocates more (once per token), but it's clearly more correct.