Replace range tables with trie for performance (#4)
Use triegen from x/text to generate tries, replacing unicode.RangeTable.

Background: unicode.Is() dominated the performance profiling. This is to be expected because the logic is mostly tight loops checking categories. It was acceptably fast; this attempts to optimize that hot path further.

The tries contain UTF8 bytes, and their payload is a bitmap of categories. Category lookups combine the trie lookup with bitwise math.

Performance improvement looks to be 30-40%. Allocations are unchanged.