Segmenter (#11)
Add new `Segmenter` type, an iterator for use with `[]byte` inputs. It is analogous to the existing `Scanner`, but doesn’t require an `io.Reader`, and is maybe 10% faster. Also adds a `SegmentAll()` convenience method, for one-liners.

Get rid of some parallelism in tests, it wasn’t buying us anything. Update docs.